Use importlib.resources to access spinegeneric/flags/*.png (#173)
* Distribute packages properly.

This wasn't working right. It would seem to work if you installed it
via `pip install -e`, but (because python's packaging system is an
inconsistent mess) but if you did, e.g.:

```
pip install git+https://github.com/spine-generic/spine-generic/
```

Then trying to use it would fail with

```
$ sg_params_checker -path-in ./

Traceback (most recent call last):

  File "/home/travis/miniconda3/bin/sg_params_checker", line 5, in <module>

    from spinegeneric.cli.params_checker import main

ModuleNotFoundError: No module named 'spinegeneric'

The command "sg_params_checker -path-in ./" exited with 1.
```

This fixes all that up.

It also *removes* the tests from packaging, by removing their __init__.py
This is on purpose! The tests can be distributed with the source code,
but users don't need them.

* Use importlib.resources to find flag files.

As of python 3.7 this is the blessed way to access associated data:
https://docs.python.org/3/library/importlib.html#module-importlib.resources

I'm bumping the min python version for this. I don't think python 3.7
should be controversial, at this point.

Co-authored-by: Julien Cohen-Adad <jcohen@polymtl.ca>