Replace importlib.resources with basic path manipulation
importlib.resources.path() is handy, but a bit overcomplicated.
It's meant to be used only on files, but here it was used to
find a folder. This is a subtle mistake, because python packages
(things you can 'import') can be a virtual union of many underlying
folders -- these are known as "namespace packages". They're neat
but they mean that a python package doesn't have a well-defined
single location in a filesystem. importlib.resources is mostly
meant to hide this and other quirks.

But we know where the test data is: it's in a subfolder of the
same folder where the tests are, so just look at the test's
__file__ and use that.

Another solution would have been to create
spinegeneric/tests/results_dummy/__init__.py, which makes results_dummy
an older-style, non-namespace package, with a single defined path
on disk, and use path_results = spinegeneric.tests.results_dummy.__file__.
But this is simpler.