Remove import path modifications
This script imports various SCT modules, so it can only be sensibly
run from an environment in which SCT is installed (unclear which
version, though). For example, this can be done as suggested in the
block comment at the top of the file:

  ${SCT_DIR}/python/envs/venv_sct/bin/python create_mosaic.py

In such a case, there's no need to modify the import path by
manually appending entries to sys.path.

This fixes most of the remaining flake8 linting errors.