Use two handlers for logging
Sending the logging output to two places (both the terminal
and a file) is exactly the kind of thing that should be
handled by the (very featureful) logging library.

As a bonus, the log messages now appear in the terminal at
the time they are generated, rather than all at the end.